% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cellAnnotation.R
\name{add_cell_annotation}
\alias{add_cell_annotation}
\title{add_cell_annotation}
\usage{
add_cell_annotation(loom, cellAnnotation)
}
\arguments{
\item{loom}{loom file handler}

\item{cellAnnotation}{data.frame with the cell information}
}
\value{
The .loom file
}
\description{
Adds the data.frame with cell information to the loom file
}
