% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bike_choice}
\alias{bike_choice}
\title{Bike Choice}
\usage{
bike_choice(dist, tripsebike)
}
\arguments{
\item{dist}{Trip Distance}

\item{tripsebike}{Probability of converting a trip into an ebike trip based on trip distance}
}
\value{
Returns a boolean: 1 for ebike and 0 for pedal bike
}
\description{
Probabilistically identifies if a trip is converted to a pedal bike trip or an ebike trip
}

