% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_rss.R
\name{calcRSS}
\alias{calcRSS}
\title{RSS}
\usage{
calcRSS(AUC, cellAnnotation, cellTypes = NULL)
}
\value{
Matrix with the regulon specificity scores
}
\description{
Calculates the regulon specificity score
}
\examples{
# TODO
}
\seealso{
The RSS was first used by Suo et al. in: 
Revealing the Critical Regulators of Cell Identity in the Mouse Cell Atlas. 
Cell Reports (2018). doi: 10.1016/j.celrep.2018.10.045
}
