% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_probability.R
\name{direct_prob_RR}
\alias{direct_prob_RR}
\title{Potential Cyclists (Relative Risk Approach)}
\usage{
direct_prob_RR(baselineSubset, DP, ebikes, equity, pcycl_baseline, region)
}
\arguments{
\item{baselineSubset}{A subset of baseline travey survey data as a data frame}

\item{DP}{Direct probability of potential cyclists}

\item{ebikes}{Boolean variable for ebikes}

\item{equity}{Boolean variable for equity (between men and women)}

\item{pcycl_baseline}{Cycling probability broken down by age and gender groups}

\item{region}{An integer variable to specify one of the nine regions of England}
}
\value{
Return IDs of all individuals satisfying \code{DP}, \code{ebikes} and \code{equity} in a specified \code{region}
}
\description{
Identifies IDs of all ppl (including also these who were cyclist already) who became potential cyclist using relative risk approach
}

