% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_probability.R
\name{directProbProportionsPPLIDs}
\alias{directProbProportionsPPLIDs}
\title{Return IDs of all ppl (including also these who were cyclist already) who became potential cyclist using proportions of cyclists group}
\usage{
directProbProportionsPPLIDs(baselineSubset, DP, ebikes, equity, pcycl_baseline,
  region)
}
\description{
Return IDs of all ppl (including also these who were cyclist already) who became potential cyclist using proportions of cyclists group
}

