% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_probability.R
\name{directProbRRPPLIDs}
\alias{directProbRRPPLIDs}
\title{Potential Cyclists (Relative Risk Approach)}
\usage{
directProbRRPPLIDs(baselineSubset, DP, ebikes, equity, pcycl_baseline, region)
}
\arguments{
\item{baselineSubset}{???}

\item{DP}{???}

\item{ebikes}{???}

\item{equity}{???}

\item{pcycl_baseline}{???}

\item{region}{???}
}
\description{
Returns IDs of all ppl (including also these who were cyclist already) who became potential cyclist using relative risk approach
}

