% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/dotHeatmap.R
=======
% Please edit documentation in R/dotheatmap.R
>>>>>>> 611f9c2f9f3102623721ea5077ed74d3c709f6a7
\name{dotHeatmap}
\alias{dotHeatmap}
\title{dotHeatmap}
\usage{
dotHeatmap(
  enrichmentDf,
  var.x = "Topic",
  var.y = "ID",
  var.col = "FC",
  col.low = "dodgerblue",
  col.mid = "floralwhite",
  col.high = "brown1",
  var.size = "p.adjust",
  min.size = 1,
  max.size = 8,
  ...
)
}
\arguments{
\item{enrichmentDf}{Input data.frame}

\item{var.x}{Variable (column) for the X axis}

\item{var.y}{Variable (column) for the Y axis}

\item{var.col}{Variable (column) that will determine the color of the dots}

\item{col.low}{Lower value color}

\item{col.mid}{Mid value color}

\item{col.high}{High value color}

\item{var.size}{Variable (column) that will determine the dot size}

\item{min.size}{Minimum dot size}

\item{max.size}{Maximum dot size}

\item{...}{Other arguments to pass to ggplot's theme()}
}
\value{
A ggplot object
}
\description{
Plots a dot-heatmap for enrichment results
}
\examples{
# TODO
}
