% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_exportsForArboreto.R
\name{exportsForArboreto}
\alias{exportsForArboreto}
\title{exportsForArboreto (alias: exportsForGRNBoost)}
\usage{
exportsForArboreto(exprMat, scenicOptions, dir = "int")
}
\arguments{
\item{exprMat}{Expression matrix}

\item{scenicOptions}{Fields used: getDatasetInfo(scenicOptions, "org"), to select the annotation database.}

\item{dir}{Output folder (by default 'int' in current working directory)}
}
\value{
Writes the output as text files.
}
\description{
Exports the TF list and expression matrix to text format to run GRNBoost (https://arboreto.readthedocs.io).
Note that it is also possible to run the full pipeline of SCENIC in python. See https://pyscenic.readthedocs.io
}
\examples{
exportsForArboreto(exprMat, scenicOptions)
}
