% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_exportsForGRNBoost.R
\name{exportsForGRNBoost}
\alias{exportsForGRNBoost}
\title{exportsForGRNBoost}
\usage{
exportsForGRNBoost(exprMat, scenicOptions)
}
\arguments{
\item{exprMat}{Expression matrix}

\item{scenicOptions}{Fields used: getDatasetInfo(scenicOptions, "org"), to select the annotation database.}
}
\value{
Writes the output as text files.
}
\description{
Exports the TF list and expression matrix to text format to run GRNBoost (https://arboretum.readthedocs.io).
}
\examples{
exportsForGRNBoost(exprMat, scenicOptions)
}
