% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_gram.R
\name{flow_gram}
\alias{flow_gram}
\title{Flowgram}
\usage{
flow_gram(baseline, MS, ebikes, equity, pcycl_baseline)
}
\arguments{
\item{baseline}{Baseline Travel Survey Dataframe}

\item{MS}{Direct probability of potential cyclists}

\item{ebikes}{Boolean variable for ebikes}

\item{equity}{Boolean variable for equity (between men and women)}

\item{pcycl_baseline}{Cycling probability broken down by age and gender groups}
}
\description{
Not sure what it does yet.
}

