% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_gram.R
\name{flowgram}
\alias{flowgram}
\title{Flowgram}
\usage{
flowgram(baseline, MS, ebikes, equity, pcycl_baseline)
}
\arguments{
\item{baseline}{???}

\item{MS}{???}

\item{ebikes}{???}

\item{equity}{???}

\item{pcycl_baseline}{???}
}
\description{
Not sure what it does yet.
}

