% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCENIC_2_createRegulons.R
\name{getDbAnnotations}
\alias{getDbAnnotations}
\title{getDbAnnotations}
\usage{
getDbAnnotations(scenicOptions)
}
\arguments{
\item{scenicOptions}{Fields used: 'scenicOptions@inputDatasetInfo$org', and 'scenicOptions@settings$db_mcVersion'}
}
\value{
The motif annotations
}
\description{
Loads the motif annotation for the given organism and database version.
}
\examples{
getDbAnnotations(scenicOptions)
}
