% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCENIC_2_createRegulons.R
\name{getDbTfs}
\alias{getDbTfs}
\title{getDbTfs}
\usage{
getDbTfs(scenicOptions)
}
\arguments{
\item{scenicOptions}{Fields used: 'scenicOptions@inputDatasetInfo$org'}
}
\value{
List of transcription factors in the databases.
}
\description{
Provides the list of transcription factors in the RcisTarget databases for the given organism
}
\examples{
getDbTfs(scenicOptions)
}
