% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonNames.getRegulonName.R
\name{getRegulonName}
\alias{getRegulonName}
\title{getRegulonName}
\usage{
getRegulonName(TFs, allRegulonNames)
}
\arguments{
\item{TFs}{...}

\item{allRegulonNames}{...}
}
\value{
...
}
\description{
TO DO
}
\details{
...
}
\seealso{
List of vignettes included in the package: \code{vignette(package="SCENIC")}
}
