% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonNames.getTF.R
\name{getTF}
\alias{getTF}
\title{getTF}
\usage{
getTF(regulonName)
}
\arguments{
\item{regulonName}{...}
}
\value{
...
}
\description{
TO DO
}
\details{
...
}
\seealso{
List of vignettes included in the package: \code{vignette(package="SCENIC")}
}
