% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expl.grnPerCell.R
\name{grnPerCell}
\alias{grnPerCell}
\title{grnPerCell}
\usage{
grnPerCell(regulonsPerGroup, binaryMat, minRegulonPerc = 0.3)
}
\arguments{
\item{regulonsPerGroup}{...}

\item{binaryMat}{...}

\item{minRegulonPerc}{...}
}
\value{
...
}
\description{
TO DO
}
\details{
...
}
\seealso{
List of vignettes included in the package: \code{vignette(package="SCENIC")}
}
