% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_importsFromPython.R
\name{importArboreto}
\alias{importArboreto}
\title{importArboreto}
\usage{
importArboreto(
  fileName,
  reorder = TRUE,
  normalizeImportance = TRUE,
  lapplyFun = BiocParallel::bplapply
)
}
\arguments{
\item{fileName}{File name of the link/adjacency list}

\item{reorder}{Whether to sort the links by decreasing weight/importance}
}
\description{
Import output from Arboreto/GRNBoost (https://arboreto.readthedocs.io)
}
\examples{
GRNBoost_linkList <- importArboreto("adjacencies.tsv")
}
\seealso{
pySCENIC (https://pyscenic.readthedocs.io)
}
