% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_exportsForGRNBoost.R
\name{importGRNBoostResults}
\alias{importGRNBoostResults}
\title{importGRNBoostResults}
\usage{
importGRNBoostResults(scenicOptions, ...)
}
\arguments{
\item{scenicOptions}{Fields used:}
}
\value{
# TODO
}
\description{
Imports and formats GRNBoost results to continue the SCENIC workflow
}
\examples{
importGRNBoostResults(scenicOptions)
}
