% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_importsFromPython.R
\name{importModulesGmt}
\alias{importModulesGmt}
\title{importModules (deprecated)}
\usage{
importModulesGmt(
  fileName,
  scenicOptions = NULL,
  saveAsDf = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{fileName}{File name of the co-expression modules (.gmt)}

\item{transpose}{saveAsDf=TRUE to save as data.frame to continue the pipeline in R (with runSCENIC_2_createRegulons)}
}
\value{
Returns the co-expression modules as list (invisible), and saves them as data.frame if requested (file name: getIntName(scenicOptions, "tfModules_asDF"))
}
\description{
Imports TF co-expression from .gmt file, as generated by pySCENIC
Deprecated: This function is no longer needed. The results can be loaded directly from the .loom file with get_regulons()
}
\examples{
pyScenicDir <- "."
tfModules <- importModules(fileName=file.path(pyScenicDir,  "modules.gmt"))
}
\seealso{
pySCENIC (https://pyscenic.readthedocs.io)
}
