% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_regulons.R
\name{onlyNonDuplicatedExtended}
\alias{onlyNonDuplicatedExtended}
\title{onlyNonDuplicatedExtended}
\usage{
onlyNonDuplicatedExtended(regulonNames)
}
\arguments{
\item{regulonNames}{Character vector containing the regulon names (e.g. rownames(AUC_))}
}
\value{
Character vector
}
\description{
Returns the regulon names filtering-out the "extended" regulons if there is a regulon based on high-confidence annotations
}
\examples{
reguNames <- c("Dlx1 (103g)", "Dlx1_extended (190g)", "Olig2_extended (29g)", "Sox9 (17g)")
onlyNonDuplicatedExtended(reguNames)
}
\seealso{
List of vignettes included in the package: \code{vignette(package="SCENIC")}
}
