% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aucTsne.R
\name{plot_aucTsne}
\alias{plot_aucTsne}
\title{...}
\usage{
plot_aucTsne(tSNE, exprMat, regulonAUC = NULL, thresholds = NULL, cex = 1,
  alphaOn = 1, alphaOff = 0.2, borderColor = adjustcolor("darkgrey", alpha
  = 0.3), offColor = "lightgray", plots = c("histogram", "binaryAUC", "AUC",
  "expression"))
}
\arguments{
\item{tSNE}{...}

\item{exprMat}{...}

\item{regulonAUC}{...}

\item{thresholds}{...}

\item{cex}{...}

\item{alphaOn}{...}

\item{alphaOff}{...}

\item{borderColor}{...}

\item{offColor}{...}

\item{plots}{...}
}
\value{
...
}
\description{
TO DO
To avoid calculating thresholds, set thresholds to FALSE
}
\details{
...
}
\seealso{
List of vignettes included in the package: \code{vignette(package="SCENIC")}
}
