% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTsne_rgb.R
\name{plotEmb_rgb}
\alias{plotEmb_rgb}
\alias{plotEmb_rgb,list-method}
\alias{plotEmb_rgb,ScenicOptions-method}
\alias{plotEmb_rgb,scenicOptions-method}
\title{plotEmb_rgb}
\usage{
plotEmb_rgb(
  object,
  regulonNames,
  aucType = "AUC",
  aucMaxContrast = 0.8,
  offColor = "#c0c0c030",
  showPlot = TRUE,
  showLegend = TRUE,
  tSNE_fileName = NULL,
  ...
)

\S4method{plotEmb_rgb}{list}(
  object,
  regulonNames,
  aucType = "AUC",
  aucMaxContrast = 0.8,
  offColor = "#c0c0c030",
  showPlot = TRUE,
  showLegend = TRUE,
  tSNE_fileName = NULL,
  ...
)

\S4method{plotEmb_rgb}{ScenicOptions}(
  object,
  regulonNames,
  aucType = "AUC",
  aucMaxContrast = 0.8,
  offColor = "#c0c0c030",
  showPlot = TRUE,
  showLegend = TRUE,
  tSNE_fileName = NULL,
  ...
)
}
\arguments{
\item{object}{If 'scenicOptions' it will use the fields: AUC matrix (continuous or binary), and default t-SNE.
If 'list', provide list(mat4col=.., emb=...)}

\item{regulonNames}{Regulons to plot}

\item{aucType}{"AUC" or "Binary"}

\item{aucMaxContrast}{To increase the AUC contrast decrease the value.}

\item{offColor}{Color por the cells completelly off. To deactivate (color as black), set to NULL.}

\item{showPlot}{Whether to plot the coloured embeddings.}

\item{showLegend}{Whether to plot add a legend to the plot.}

\item{tSNE_fileName}{tSNE file name. If null, the default t-SNE is used.}

\item{...}{Other arguments to pass to the \code{plot} function.}
}
\value{
The cell colors (invisible)
}
\description{
Colors the embeddings (t-SNE/Umap) based on the activity of 3 (groups of) regulons
}
\examples{
par(mfrow=c(1,2))

regulonNames <- c("Dlx1", "Sox8")
cellCol <- plotEmb_rgb(scenicOptions, regulonNames, aucType="AUC", aucMaxContrast=0.6)
text(-5,-23, attr(cellCol,"red"), col="red", cex=.7)
text(-10,-18, attr(cellCol,"green"), col="green", cex=.7)

regulonNames <- list(red=c("Dlx1","Dlx5"),
                     green=c("Neurod1"),
                     blue=c("Sox8"))
cellCol <- plotEmb_rgb(scenicOptions, regulonNames, aucType="Binary")

#OR if already loaded:
mat2col <- readRDS("int/2.2_motifs_AUC.Rds")
emb <- readRDS("int/tSNE_AUC_05pcs_05perpl.Rds")
cellCol <- plotEmb_rgb(list(mat2col=mat2col, emb=emb), regulonNames, aucType="Binary")

}
