% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_rss.R
\name{plotRSS}
\alias{plotRSS}
\title{Plot RSS}
\usage{
plotRSS(
  rss,
  labelsToDiscard = NULL,
  zThreshold = 1,
  cluster_columns = FALSE,
  order_rows = TRUE,
  thr = 0.01,
  varName = "cellType",
  col.low = "grey90",
  col.mid = "darkolivegreen3",
  col.high = "darkgreen",
  revCol = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{rss}{Output of calcRSS()}

\item{labelsToDiscard}{Cell labels to ignore (i.e. do not plot). IMPORTANT: All cells in the analysis should be included when calculating the RSS.}

\item{zThreshold}{}

\item{cluster_columns}{}

\item{order_rows}{}

\item{thr}{}

\item{varName}{}

\item{col.low}{}

\item{col.mid}{}

\item{col.high}{}

\item{verbose}{}

\item{setName}{Gene-set or cell type name to plot with plotRSS_oneSet()}

\item{n}{Number of top regulons to label in plotRSS_oneSet(). Default: 5}
}
\value{
Matrix with the regulon specificity scores
}
\description{
Plots an overview of the regulon specificity score
}
\examples{
# TODO
}
