% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTsne_AUCellApp.R
\name{plotTsne_AUCellApp}
\alias{plotTsne_AUCellApp}
\title{plotTsne_AUCellApp}
\usage{
plotTsne_AUCellApp(
  scenicOptions,
  exprMat,
  tSNE_fileName = NULL,
  skipDuplicatedExtended = TRUE
)
}
\arguments{
\item{scenicOptions}{Fields used: aucell_regulonAUC, aucell_thresholds, cellInfo, and default tSNE}

\item{exprMat}{Expression matrix}

\item{tSNE_fileName}{tSNE file name. If null, the default t-SNE is used.}

\item{skipDuplicatedExtended}{Skip extended regulons if there is a high-confidence regulon for the same TF}
}
\value{
Invisible: aucellApp
}
\description{
Creates the AUCell shiny app to visualize the regulon activities on the t-SNE
}
\examples{
logMat <- exprMat # Better if it is logged/normalized
aucellApp <- plotTsne_AUCellApp(scenicOptions, logMat)
savedSelections <- shiny::runApp(aucellApp)
}
