% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTsne_AUCellHtml.R
\name{plotTsne_AUCellHtml}
\alias{plotTsne_AUCellHtml}
\title{plotTsne_AUCellHtml}
\usage{
plotTsne_AUCellHtml(scenicOptions, exprMat, fileName, tSNE = NULL)
}
\arguments{
\item{scenicOptions}{Fields used: aucell_regulonAUC, aucell_thresholds (int), and if \code{tSNE=NULL} the default tSNE.}

\item{exprMat}{Expression matrix (if NULL, TF expression will not be plotted)}

\item{fileName}{Location for the plots.}

\item{tSNE}{If null, it plots the default t-SNE.}
}
\value{
NULL. The plots are saved as HTML with the name stored in \code{fileName}. 
If you are using RStudio, you might need to download the .html file and plots folder to view locally.
}
\description{
Wrapper for the 'AUCell_plotTSNE' function: plots the regulon activities on the t-SNE.
}
\examples{
plotTsne_AUCellHtml(scenicOptions, exprMat, fileName)
}
