% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTsne_compareSettings.R
\name{plotTsne_compareSettings}
\alias{plotTsne_compareSettings}
\title{plotTsne_compareSettings}
\usage{
plotTsne_compareSettings(fileNames, scenicOptions, varName = NULL, ...)
}
\arguments{
\item{fileNames}{t-SNE files to plot}

\item{scenicOptions}{imports cellInfo and colVars}

\item{varName}{If a varName is given, it plots all the t-SNEs (not saved to file), colouring only for that variable.
If null, it plots all the cellInfo colums for all the t-SNEs (each t-SNE in an individual file). 
To plot all cellInfo colums, but as individual plots, use \code{TRUE}.}

\item{...}{Other arguments to pass to \code{AUCell::plotTsne_cellProps}}
}
\value{
Plots
}
\description{
Plots several t-SNEs coloured by cell properties (to compare different t-SNE settings)
}
\examples{
fileNames <- paste0("int/", grep("tSNE_", list.files("int"), value=T))
par(mfrow=c(2,2))
plotTsne_compareSettings(fileNames, scenicOptions, varName="CellType")
}
