% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTsne_AUCell.R
\name{plotTsne_regulonActivityHTML}
\alias{plotTsne_regulonActivityHTML}
\title{plotTsne_AUCell}
\usage{
plotTsne_regulonActivityHTML(scenicOptions, exprMat, fileName, tSNE = NULL)
}
\arguments{
\item{scenicOptions}{Fields used: aucell_regulonAUC, aucell_thresholds (int), s3_AUCtSNE_colAct (output), and default tsne.}

\item{exprMat}{Expression matrix}

\item{fileName}{Location for the plots.}

\item{tSNE}{If null, it plots the default t-SNE.}
}
\value{
NULL. The plots are saved as HTML with the name stored in \code{getOutName(scenicOptions, "s3_AUCtSNE_colAct")}. 
If you are using RStudio, you might need to download the .html file and plots folder to view locally.
}
\description{
Wrapper for the 'AUCell_plotTSNE' function: plots the regulon activities on the t-SNE
}
\examples{
plotTsne_AUCell(scenicOptions, exprMat)
}
