% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTsne_rgb.R
\name{plotTsne_rgb}
\alias{plotTsne_rgb}
\title{plotTsne_rgb}
\usage{
plotTsne_rgb(
  scenicOptions,
  regulonNames,
  aucType = "AUC",
  aucMaxContrast = 0.8,
  offColor = "#c0c0c030",
  showPlot = TRUE,
  showLegend = TRUE,
  tSNE_fileName = NULL,
  ...
)
}
\arguments{
\item{scenicOptions}{Fields used: AUC matrix (continuous or binary), default t-SNE.}

\item{regulonNames}{Regulons to plot}

\item{aucType}{"AUC" or "Binary"}

\item{aucMaxContrast}{To increase the AUC contrast decrease the value.}

\item{offColor}{Color por the cells completelly off. To deactivate (color as black), set to NULL.}

\item{showPlot}{Whether to plot add a legend to the plot.}

\item{tSNE_fileName}{tSNE file name. If null, the default t-SNE is used.}

\item{...}{Other arguments to pass to the \code{plot} function.}
}
\value{
The cell colors (invisible)
}
\description{
Colors the t-SNE based on the activity of 3 (groups of) regulons
}
\examples{
par(mfrow=c(1,2))

regulonNames <- c("Dlx1", "Sox8")
cellCol <- plotTsne_rgb(scenicOptions, regulonNames, aucType="AUC", aucMaxContrast=0.6)
text(-5,-23, attr(cellCol,"red"), col="red", cex=.7)
text(-10,-18, attr(cellCol,"green"), col="green", cex=.7)

regulonNames <- list(red=c("Dlx1","Dlx5"),
                     green=c("Neurod1"),
                     blue=c("Sox8"))
cellCol <- plotTsne_rgb(scenicOptions, regulonNames, aucType="Binary")

}
