% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prob_cycling}
\alias{prob_cycling}
\title{Calculates probability of a given trip being cycled}
\usage{
prob_cycling(age, sex, dist, ebikes, pcycl_baseline)
}
\arguments{
\item{age}{Either 16.59 or 60plus}

\item{sex}{Male or Female}

\item{dist}{Distance of the trip}

\item{ebikes}{Boolean variable if it's an ebike scenario or not}

\item{pcycl_baseline}{Table (as a data frame) with age and gender specific probability of cycling a trip}
}
\value{
Probability of a cyle trip
}
\description{
Calculates probability of a given trip being cycled
}

