% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prob_to_odds}
\alias{prob_to_odds}
\title{Probability to Odds ratio}
\usage{
prob_to_odds(x)
}
\arguments{
\item{x}{Odds ratio}
}
\value{
Probability
}
\description{
Converts probability to odds ratio
}

