% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_regulons.R
\name{regulon_plotExpression}
\alias{regulon_plotExpression}
\title{regulon_plotExpression}
\usage{
regulon_plotExpression(
  exprMat,
  regulonsSelected,
  nCells = 500,
  cellInfo = NULL,
  colVars = NULL,
  color = c("black", "goldenrod", "yellow"),
  ...
)
}
\arguments{
\item{exprMat}{Expression matrix}

\item{regulonsSelected}{Regulons to plot}

\item{nCells}{Number of cells to plot (random subset of cells for faster plotting)}

\item{cellInfo}{Cell labels to plot in the heatmap (data.frame)}

\item{colVars}{Colors for the cell labels}

\item{color}{Expression color range}

\item{...}{Other arguments to pass to \code{NMF::aheatmap}}
}
\value{
Plots the heatmap
}
\description{
Plots the expression of the genes in the regulon (as heatmap)
}
\examples{
loomPath <- system.file(package="SCENIC", "examples/mouseBrain_toy.loom")
exprMat <- SCopeLoomR::get_dgem(SCopeLoomR::open_loom(loomPath)) # normalized preferred
dim(exprMat)

cellInfo <- data.frame(loadFile(scenicOptions, getDatasetInfo(scenicOptions, "cellInfo")))
colVars <- loadFile(scenicOptions, getDatasetInfo(scenicOptions, "colVars"))

exprMat_log <- log2(exprMat+1)
regulons <- loadInt(scenicOptions, "regulons")
regulonNames <- getRegulonName(c("Dlx1") , names(regulons))
regulon_plotExpression(exprMat_log, regulons[regulonNames], cellInfo=cellInfo, colVars=colVars) #, filename="regulonExpression.pdf")
}
