% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_regulons.R
\name{regulonsToGeneLists}
\alias{regulonsToGeneLists}
\title{regulonsToGeneSet}
\usage{
regulonsToGeneLists(incidMat)
}
\arguments{
\item{incidMat}{Incidence matrix with TFs as rows and GENES as columns}
}
\value{
List of genes per regulon
}
\description{
Converts the regulons stored as incidence matrix to gene sets (e.g. list of genes)
}
\examples{
regulonsToGeneSet(incidMat)
}
