% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCENIC.R
\name{runSCENIC}
\alias{runSCENIC}
\title{Run SCENIC (steps 1.2 trhough 4)}
\usage{
runSCENIC(exprMat = NULL, org = NULL, cellInfo = NULL, colVars = NULL,
  stepsToRun = c("1.2", "2", "3.1", "3.2", "4"), nCores = 4, seed = 123,
  verbose = TRUE, showAlternativeTsnes = TRUE)
}
\arguments{
\item{exprMat}{expression matrix}

\item{org}{"hg19" for human, "mm9" for mouse, or "dm6" for drosophila melanogaster. Determines the databases used for RcisTarget.}

\item{cellInfo}{Phenodata information for the cells}

\item{colVars}{Colors to use to plot the cellInfo (formatted as a list, same format as in NMF::aheatmap)}

\item{stepsToRun}{Determines the Steps of SCENIC to run (matches the numbers/order in the title of the tutorials). By default: c("1.2", "2", "3.1", "3.2", "4").}

\item{nCores}{Number of cores to use for computation}

\item{seed}{Seed to use for "random" computations (for reproducibility of the results)}

\item{verbose}{TRUE/FALSE. Whether to show progress messages on screen. If false, the progress messages will be saved in the log file: output/runSCENIC_log.txt}

\item{showAlternativeTsnes}{TRUE/FALSE. Whether to plot t-SNEs with diferent parameters.
In general it is recommended to plot multiple t-SNEs to evaluate the stability of the results, but it might take long time to run. For big datasets, it might be better to manually choose/try the parameters following the individual tutorials.}
}
\value{
No value is returned. The outputs of this function are saved in the folder "output" the current working directory. Other files are saved in the folder "int" (e.g. for further exploration or to perform alternative analyses).
}
\description{
Runs SCENIC steps after GENIE3.
This function should be equivalent to running the following tutorials sequentially :
- Step1.2_Regulons.Rmd:  Identifying regulons (direct TF targets) based on DNA motif enrichment.
- Step3.1_NwActivity.Rmd: Analyzing the network activity in each individual cell (part 1: calculate AUC).
- Step3.2_BinaryNwActivity.Rmd: Creating the binary activity matrix (convert the network activity into ON/OFF).
- Step4_Clustering.Rmd: Identify stable cell states based on their gene regulatory network activity (cell clustering)
(e.g. all steps and parameters are the same)
}
\details{
...
}
\examples{
# For a detailed example see the package vignette:
vignette("SCENIC_runningStep1nWrapper")


\dontrun{

setwd("SCENIC_MouseBrain")
load("data/esetMouseBrain.RData")
exprMat <- exprs(esetMouseBrain)
# Optional: add log for TF expression plot, it does not affect any other calculation
exprMat <- log2(exprMat+1)
dim(exprMat)

load("data/colVars.RData")
cellInfo <- pData(esetMouseBrain)[colnames(exprMat), names(colVars), drop=F]

library(SCENIC)
runSCENIC(exprMat=exprMat, org="mm9", cellInfo=cellInfo, colVars=colVars, nCores=4,  stepsToRun=c("1.2", "2", "3.1", "3.2", "4"))

}
}
\seealso{
List of vignettes included in the package: \code{vignette(package="SCENIC")}
}
