% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCENIC_3_scoreCells.R
\name{runSCENIC_3_scoreCells}
\alias{runSCENIC_3_scoreCells}
\title{runSCENIC_3_scoreCells}
\usage{
runSCENIC_3_scoreCells(
  scenicOptions,
  exprMat,
  skipBinaryThresholds = FALSE,
  skipHeatmap = FALSE,
  skipTsne = FALSE
)
}
\arguments{
\item{scenicOptions}{Fields used: TODO}

\item{exprMat}{Expression matrix}

\item{skipBinaryThresholds}{Whether to skip the automatic binarization step}

\item{skipHeatmap}{Whether to plot the AUC heatmap}

\item{skipTsne}{Whether to plot the t-SNE}
}
\value{
The output is written in the folders 'int' and 'ouput'
}
\description{
Step 3: AUCell (scoring the regulons on the individual cells)
}
\details{
See the detailed vignette explaining the internal steps.
}
\examples{
runSCENIC_3_scoreCells(scenicOptions)
}
