% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCENIC_4.R
\name{runSCENIC_4_aucell_binarize}
\alias{runSCENIC_4_aucell_binarize}
\title{runSCENIC_4_aucell_binarize}
\usage{
runSCENIC_4_aucell_binarize(
  scenicOptions,
  skipBoxplot = FALSE,
  skipHeatmaps = FALSE,
  skipTsne = FALSE,
  exprMat = NULL
)
}
\arguments{
\item{scenicOptions}{Fields used: TODO}

\item{skipBoxplot}{Whether to plot the boxplots}

\item{skipHeatmaps}{Whether to plot the Binary heatmaps}

\item{skipTsne}{Whether to calculate the binary t-SNE (and optionally, generate a HTML preview)}

\item{exprMat}{If !skipTsne, the expression matrix can be provided to plot the TF expression on the t-SNE (plotTsne_AUCellHtml)}
}
\value{
The output is written in the folders 'int' and 'ouput'
}
\description{
Step 4: Binarize the AUC (and, optional: re-cluster)
}
\details{
See the detailed vignette explaining the internal steps.
}
\examples{
runSCENIC_4_aucell_binarize(scenicOptions)
}
