% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageInfo.R
\docType{package}
\name{scenic-package}
\alias{scenic-package}
\alias{scenic}
\title{Scenario Generation Tool (scenic)}
\description{
Description: An implementation of scenario generation based on
travel survey data. It takes into account variable names and type
of scenarios users are interested in. See the ICT website for more
information, documentation and examples.
}
\examples{

}
\references{
\url{http://www.cedar.iph.cam.ac.uk/research/modelling/ithim/}, \url{https://ithim.ghi.wisc.edu/}
}
\author{
Ali Abbas \email{m.ali.abbas@gmal.com}
}
