% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ScenicOptions _ doc.R
\docType{class}
\name{ScenicOptions-class}
\alias{ScenicOptions-class}
\alias{ScenicOptions}
\alias{loadInt}
\alias{show}
\alias{show,ScenicOptions-method}
\alias{getDatasetInfo}
\alias{getDatasetInfo,ScenicOptions-method}
\alias{getDatabases}
\alias{getDatabases,ScenicOptions-method}
\alias{getSettings}
\alias{getSettings,ScenicOptions-method}
\alias{getOutName}
\alias{getOutName,ScenicOptions-method}
\alias{getIntName}
\alias{getIntName,ScenicOptions-method}
\alias{loadFile}
\alias{loadFile,ScenicOptions-method}
\alias{loadInt}
\alias{loadInt,ScenicOptions-method}
\title{Object to store SCENIC settings}
\usage{
\S4method{show}{ScenicOptions}(object)

getDatasetInfo(object, ...)

\S4method{getDatasetInfo}{ScenicOptions}(object, slotName = NULL)

getDatabases(object, ...)

\S4method{getDatabases}{ScenicOptions}(object)

getSettings(object, ...)

\S4method{getSettings}{ScenicOptions}(object, slotName = NULL)

getOutName(object, ...)

\S4method{getOutName}{ScenicOptions}(object, out_type = NULL)

getIntName(object, ...)

\S4method{getIntName}{ScenicOptions}(object, int_type = NULL)

loadFile(object, ...)

\S4method{loadFile}{ScenicOptions}(object, fileName, verbose = FALSE,
  ifNotExists = c("error", "null"), ...)

loadInt(object, ...)

\S4method{loadInt}{ScenicOptions}(object, int_type = NULL, ...)
}
\arguments{
\item{object}{}
}
\value{
\itemize{
\item show: Prints a summary of the object
\item loadInt: Loads the selected object
}
}
\description{

}
