% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expl.selectCells.R
\name{selectCells}
\alias{selectCells}
\title{selectCells}
\usage{
selectCells(tSNE_binary, binaryRegulonActivity = NULL, verbose = TRUE)
}
\arguments{
\item{tSNE_binary}{...}

\item{binaryRegulonActivity}{...}

\item{verbose}{...}
}
\value{
...
}
\description{
TO DO
}
\details{
...
}
\seealso{
List of vignettes included in the package: \code{vignette(package="SCENIC")}
}
