% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trip_speed}
\alias{trip_speed}
\title{Trip Speed}
\usage{
trip_speed(age, sex, bike_type)
}
\arguments{
\item{age}{Individual age: 16.59 or 60plus}

\item{sex}{Male or Female}

\item{bike_type}{0 for pedal and 1 for ebike}
}
\value{
Trip Speed based on age, sex and bike_type
}
\description{
Calculates speed of a cycling trip, based on following parameters:\cr
1. age: With two categories 16-59 and 60+ year olds\cr
2. gender: Male or Female\cr
3. bike_type: A boolean variable whether it's a pedal bike (0) or an ebike (1)\cr
}

