% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsneAUC.R
\name{tsneAUC}
\alias{tsneAUC}
\title{tsneAUC}
\usage{
tsneAUC(
  scenicOptions,
  aucType = NULL,
  nPcs = NULL,
  perpl = NULL,
  filePrefix = NULL,
  seed = NULL,
  onlyHighConf = FALSE,
  ...
)
}
\arguments{
\item{scenicOptions}{Fields used:}

\item{aucType}{"AUC" for continuous AUC, or "Binary" . By default it uses the value in \code{getSettings(scenicOptions,"defaultTsne/dims")}}

\item{nPcs}{Number of PCs to use for the t-SNE (can be several values, see examples). By default it uses the value in \code{getSettings(scenicOptions,"defaultTsne/dims")}}

\item{perpl}{Perplexity for the t-SNE (can be several values, see examples). By default it uses the value in \code{getSettings(scenicOptions,"aucell/tsne_perpl")}}

\item{filePrefix}{Prefix for the file to save the t-SNE as. It is saved with the following format: "PREFIX_aucType_nPcs_perpl.Rds"}

\item{seed}{Seed for the t-SNEs. By default it uses the value in \code{getSettings(scenicOptions,"seed")}}

\item{onlyHighConf}{By default (TRUE) it only uses only regulons based on high-confidence annotations (i.e. "extended" are skipped)}

\item{...}{Other arguments to pass to the \code{Rtsne::Rtsne} function.}
}
\value{
The file name(s) in whihc the tSNE is saved.
}
\description{
Calculates the t-SNE based on the regulon activity. (i.e. binary/continuous AUC)
}
\examples{
## TO DO (See the vignette)
# tSNE <- readRDS(tsneFileName(scenicOptions))
# tsneAUC(scenicOptions)
}
