% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_viewMotifs.R
\name{viewMotifs}
\alias{viewMotifs}
\title{viewMotifs}
\usage{
viewMotifs(
  tableSubset,
  motifCol = c("motif", "bestMotif", "MotifID"),
  dbVersion = "v9",
  nSignif = 3,
  colsToShow = c(motifEnrichment = c("motifDb", "logo", "NES", "geneSet",
    "TF_highConf"), regulonTargets = c("TF", "gene", "nMotifs", "bestMotif", "logo",
    "NES", "highConfAnnot", "Genie3Weight")),
  options = list(pageLength = 50),
  ...
)
}
\arguments{
\item{tableSubset}{Motif information table to show}

\item{motifCol}{Column name containing the motif logo ID (to pass to RcisTarget::addLogo). If NULL the logo is not added.}

\item{dbVersion}{Database version (to pass to RcisTarget::addLogo).}

\item{nSignif}{Number of significant digits to show for numeric columns}

\item{colsToShow}{(No warning is shown for missing columns)}

\item{options}{argument to pass to DT::datatable()}

\item{...}{Any other other arguments to pass to DT::datatable()}
}
\value{
Writes the output in the file name stored in: \code{getIntName(scenicOptions, "corrMat")}
}
\description{
Shows the motif information table as HTML using DT::datatable()
}
\examples{
library(SCENIC)
scenicOptions <- readRDS("int/scenicOptions.Rds")

regulonTargetsInfo <- loadInt(scenicOptions, "regulonTargetsInfo")
tableSubset <- regulonTargetsInfo[TF=="Stat6" & highConfAnnot==TRUE]
viewMotifs(tableSubset)

motifEnrichment_selfMotifs_wGenes <- loadInt(scenicOptions, "motifEnrichment_selfMotifs_wGenes")
tableSubset <- motifEnrichment_selfMotifs_wGenes[highlightedTFs=="Sox8"]
viewMotifs(tableSubset)
}
\seealso{
\code{?DT::datatable()}
}
