% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_probability.R
\name{directProbProportionsPPLIDs}
\alias{directProbProportionsPPLIDs}
\title{Potential Cyclists (Proportions Approach)}
\usage{
directProbProportionsPPLIDs(baselineSubset, DP, ebikes, equity, pcycl_baseline,
  region)
}
\arguments{
\item{baselineSubset}{???}

\item{DP}{???}

\item{ebikes}{???}

\item{equity}{???}

\item{pcycl_baseline}{???}

\item{region}{???}
}
\description{
Returns IDs of all ppl (including also these who were cyclist
already) who became potential cyclist using proportions approach
}

