% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_gram.R
\name{flow_gram}
\alias{flow_gram}
\title{Flow_gram}
\usage{
flow_gram(baseline, DP, ebikes, equity, pcycl_baseline)
}
\arguments{
\item{baseline}{Baseline Travel Survey Dataframe}

\item{DP}{Direct probability of potential cyclists}

\item{ebikes}{Boolean variable for ebikes}

\item{equity}{Boolean variable for equity (between men and women)}

\item{pcycl_baseline}{Cycling probability broken down by age and gender groups}
}
\description{
Creates a scenario based on three paramets:\cr 
1. Direct Probability (Percentage of potential cyclists in the overall population). Its value is in percentages (between 0 and 1)\cr
2. Ebikes: A boolean variable indicating whether ebikes are being used or not. This allows longer journeys to be converted into ebike journeys.\cr
3. Equity: A boolean variable whether women have the same probablity of converting their trips into cycling trips.\cr
}
\details{
At the end, this function returns a data frame (similar to the baseline data frame), with new cycling trips
}

