% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_geneFiltering.R
\name{geneFiltering}
\alias{geneFiltering}
\title{geneFiltering}
\usage{
geneFiltering(
  exprMat,
  scenicOptions,
  minCountsPerGene = 3 * 0.01 * ncol(exprMat),
  minSamples = ncol(exprMat) * 0.01
)
}
\arguments{
\item{exprMat}{Expression matrix}

\item{scenicOptions}{scenicOptions object (slots used: RcisTarget databases and genesKept)}

\item{minCountsPerGene}{Minimum counts per gene required}

\item{minSamples}{Minimum number of samples (cells) in which the gene should be detected}
}
\description{
Gene filtering
}
\examples{
setwd("SCENIC_MouseBrain")
scenicOptions <- readRDS("int/scenicOptions.Rds")

loomPath <- system.file(package="SCENIC", "examples/mouseBrain_toy.loom")
exprMat <- SCopeLoomR::get_dgem(SCopeLoomR::open_loom(loomPath))

genesKept <- geneFiltering(exprMat, scenicOptions=scenicOptions,
                           minCountsPerGene=3*.01*ncol(exprMat),
                           minSamples=ncol(exprMat)*.01)
overrideOptions <- list(dbFilePath="databases/mm9-500bp-upstream-7species.mc9nr.feather", outFile_genesKept=NULL)
}
