% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonNames.getTF.R
\name{getTF}
\alias{getTF}
\title{getTF}
\usage{
getTF(regulonName)
}
\arguments{
\item{regulonName}{Character vector containing regulon names}
}
\value{
Character vector containing the TF (including the "_extended" sufix if appropiate)
}
\description{
Returns the TF associated to a regulon name (e.g. removes the #genes sufix)
}
\details{
...
}
\seealso{
List of vignettes included in the package: \code{vignette(package="SCENIC")}
}
