% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_regulons.R
\name{getTF}
\alias{getTF}
\title{getTF}
\usage{
getTF(regulonName, sep = "\\\\s")
}
\arguments{
\item{regulonName}{Character vector containing regulon names}

\item{sep}{Character used to separate the TF name (as regexpr), if different from space.}
}
\value{
Character vector containing the TF (including the "_extended" sufix if appropiate)
}
\description{
Returns the TF associated to a regulon name (e.g. removes the #genes sufix)
}
\details{
...
}
\examples{
reguNames <- c("Dlx1 (103g)", "Dlx1_extended (190g)", "Olig2_extended (29g)", "Sox9 (17g)")
getTF(reguNames)
}
\seealso{
List of vignettes included in the package: \code{vignette(package="SCENIC")}
}
