% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_importsFromPython.R
\name{importAUCfromText}
\alias{importAUCfromText}
\title{importAUCfromText (deprecated)}
\usage{
importAUCfromText(
  fileName,
  rows = "regulons",
  columns = "cells",
  transpose = TRUE,
  newNames = NULL
)
}
\arguments{
\item{fileName}{File name of the AUC scores (transposed: gene-sets in rows, cells in columns)}

\item{rows}{Type of data stored as rows (only for informative purposes) Default: "regulons"}

\item{columns}{Type of data stored as columns (only for informative purposes) Default: "cells"}

\item{transpose}{Whether to transpose the input matrix}
}
\description{
Import AUCell matrix from text format, as generated by pySCENIC
Deprecated: This function is no longer needed. The results can be loaded directly from the .loom file with get_regulons_AUC()
}
\examples{
regulonAUC <- importAUCfromText("aucMatrix.tsv")
}
\seealso{
pySCENIC (https://pyscenic.readthedocs.io)
}
