% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonNames.onlyNonDirectExtended.R
\name{onlyNonDirectExtended}
\alias{onlyNonDirectExtended}
\title{onlyNonDirectExtended}
\usage{
onlyNonDirectExtended(regulonNames)
}
\arguments{
\item{regulonNames}{...}
}
\value{
...
}
\description{
TO DO
}
\details{
...
}
\seealso{
List of vignettes included in the package: \code{vignette(package="SCENIC")}
}
