% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGenie3.R
\name{runGenie3}
\alias{runGenie3}
\title{runGenie3}
\usage{
runGenie3(
  exprMat,
  scenicOptions,
  nParts = 10,
  resumePreviousRun = FALSE,
  allTFs = getDbTfs(scenicOptions),
  ...
)
}
\arguments{
\item{exprMat}{Expression matrix}

\item{scenicOptions}{Fields used: genie3wm, genie3ll (int)}

\item{nParts}{Number of pieces to fragment the partial results into (to check progress or in case the job gets interrupted). Default: 10}

\item{resumePreviousRun}{Reload partial results from a previous (interrupted) run and resume execution (default: FALSE).}

\item{allTFs}{List of all TFs for the organism. The default value is all genes annotated as 'TFs' in the database, which is appropriate for most analyses.
(This argument allows to modify this list, for example removing genes that might be wrongly annotated as TF in some organisms/database versions).
Note that selecting only a small subset of TFs for the analysis will modify the output, and therefore the interpretation of the results!}

\item{...}{Other arguments to pass to GENIE3 (e.g. ntrees=500)}
}
\value{
Writes the output as \code{getIntName(scenicOptions, "genie3ll")}
}
\description{
Runs GENIE3 and formats the output
}
\examples{
library(SCENIC)
scenicOptions <- readRDS("int/scenicOptions.Rds")

loomPath <- system.file(package="SCENIC", "examples/mouseBrain_toy.loom")
exprMat <- SCopeLoomR::get_dgem(SCopeLoomR::open_loom(loomPath))

genesKept <- loadInt(scenicOptions, "genesKept")
exprMatrix_filtered <- exprMat[genesKept,]
exprMat_filtered <- log2(exprMatrix_filtered+1) 

runGenie3(exprMat_filtered, scenicOptions)
}
