% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCENIC_1_coexNetwork2modules.R
\name{runSCENIC_1_coexNetwork2modules}
\alias{runSCENIC_1_coexNetwork2modules}
\title{runSCENIC_1_coexNetwork2modules}
\usage{
runSCENIC_1_coexNetwork2modules(
  scenicOptions,
  weightThreshold = getSettings(scenicOptions, "modules/weightThreshold"),
  topThr = c(0.005),
  nTopTfs = c(5, 10, 50),
  nTopTargets = 50,
  aFun = topPerTf,
  corrThr = 0.03,
  linkList = NULL,
  corrMat = NULL,
  weightCol = "weight",
  verbose = getSettings(scenicOptions, "verbose")
)
}
\arguments{
\item{scenicOptions}{Fields used: TODO}

\item{weightThreshold}{Global weight threshold. Links with lower weight will be ignored in ALL modules.}
}
\value{
The output is written in the folders 'int' and 'ouput'
}
\description{
Step 1: Convert the output from GENIE3/GRNBoost to co-expression modules
}
\details{
See the detailed vignette explaining the internal steps.
}
\examples{
runSCENIC_1_coexNetwork2modules(scenicOptions)
}
