% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCENIC_2_createRegulons.R
\name{runSCENIC_2_createRegulons}
\alias{runSCENIC_2_createRegulons}
\title{runSCENIC_2_createRegulons}
\usage{
runSCENIC_2_createRegulons(
  scenicOptions,
  minGenes = 20,
  coexMethods = NULL,
  minJakkardInd = 0.8,
  signifGenesMethod = "aprox",
  onlyPositiveCorr = TRUE,
  onlyBestGsPerMotif = TRUE,
  dbIndexCol = "features"
)
}
\arguments{
\item{scenicOptions}{Fields used: TODO}

\item{minGenes}{Minimum size of co-expression gene set (default: 20 genes)}

\item{coexMethods}{Allows to select the method(s) used to generate the co-expression modules}

\item{minJakkardInd}{Merge overlapping modules (with Jakkard index >=minJakkardInd; reduces running time).}

\item{signifGenesMethod}{Method for 'addSignificantGenes'}

\item{onlyPositiveCorr}{Whether to include only positive-correlated targets in the regulons (default: TRUE).}

\item{dbIndexCol}{Column containing the feature name in the database (default: 'features')}
}
\value{
The output is written in the folders 'int' and 'ouput'
}
\description{
Step 2: RcisTarget (prune co-expression modules using TF-motif enrichment analysis)
}
\details{
See the detailed vignette explaining the internal steps.
}
\examples{
scenicOptions <- readRDS("int/scenicOptions.Rds")
# In case any settings need to be modified:
scenicOptions@settings$nCores <- 20
scenicOptions@inputDatasetInfo$org <- "mgi" 

runSCENIC_2_createRegulons(scenicOptions)
}
