% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_regulons.R
\name{selectRegulons}
\alias{selectRegulons}
\title{selectRegulons}
\usage{
selectRegulons(regulons, tfs, onlyNonDuplicatedExtended = FALSE)
}
\arguments{
\item{regulons}{Regulons (list) or regulonAUC object}

\item{tfs}{TFs to select}

\item{onlyNonDuplicatedExtended}{Whether to filter with the function 'onlyNonDuplicatedExtended'}
}
\value{
The selected regulons
}
\description{
Selects the regulons for the given TFs
}
\examples{
selectRegulons(regulons, c("Dlx5", "Olig2"))
selectRegulons(regulonAUC, "Tef")
}
