% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trip_speed}
\alias{trip_speed}
\title{Trip Speed}
\usage{
trip_speed(age, sex, bike_type)
}
\arguments{
\item{baseline}{Baseline Travel Survey Dataframe}

\item{DP}{Direct probability of potential cyclists}

\item{ebikes}{Boolean variable for ebikes}

\item{equity}{Boolean variable for equity (between men and women)}

\item{pcycl_baseline}{Cycling probability broken down by age and gender groups}
}
\description{
Returns speed of a cycling trip, based on following parameters:
1. age: With two categories 16-59 and 60+ year olds
2. gender: Male or Female
3. bike_type: A boolean variable whether it's a pedal bike (0) or an ebike (1)
}

